requirements.txt
botbuilder-integration-aiohttp>=4.14.0
openai==0.26.5
